Move queue into a shared directory.
