+++ lsif/src/server/routes/lsif.ts
@@ -9,7 +9,7 @@ import { addTags, logAndTraceCall, TracingContext } from '../../shared/tracing'
 import { Backend, ReferencePaginationCursor } from '../backend/backend'
 import { DEFAULT_REFERENCES_NUM_REMOTE_DUMPS, STORAGE_ROOT } from '../settings'
 import { encodeCursor, parseCursor } from '../pagination/cursor'
-import { enqueue } from '../../queue'
+import { enqueue } from '../../shared/queue/queue'
 import { extractLimit } from '../pagination/limit-offset'
 import { Logger } from 'winston'
 import { nextLink } from '../pagination/link'
