+++ lsif/src/server/server.ts
@@ -13,7 +13,7 @@ import { createLogger } from '../shared/logging'
 import { createLsifRouter } from './routes/lsif'
 import { createMetaRouter } from './routes/meta'
 import { createPostgresConnection } from '../shared/database/postgres'
-import { createQueue, ensureOnlyRepeatableJob } from '../queue'
+import { createQueue, ensureOnlyRepeatableJob } from '../shared/queue/queue'
 import { createTracer } from '../shared/tracing'
 import { dbFilename, dbFilenameOld, ensureDirectory } from '../util'
 import { default as tracingMiddleware } from 'express-opentracing'
