+++ lsif/src/worker/worker.ts
@@ -7,7 +7,7 @@ import { createCleanOldJobsProcessor } from './processors/clean-old-jobs'
 import { createConvertJobProcessor } from './processors/convert'
 import { createLogger } from '../shared/logging'
 import { createPostgresConnection } from '../shared/database/postgres'
-import { createQueue } from '../queue'
+import { createQueue } from '../shared/queue/queue'
 import { createUpdateTipsJobProcessor } from './processors/update-tips'
 import { ensureDirectory } from '../util'
 import { followsFrom, FORMAT_TEXT_MAP, Span, Tracer } from 'opentracing'
